Add functionality to Activity & Question entities, remove redundance
Added functionality (methods for inserting, deleting, and getting elements from the database) to the Activity.java file, as well as documentation to both Activity.java and Question.java. 

Changed from PathVariable to PathParam. 

Deleted the Activity.java and Question.java files from the database folder, as they are now in the commons/game folder among the other game mechanics.