There is now only one injector.
When the app is started the MyFXML object is passed to the main controller. The main controller IS NOT used by any scenes anymore and is instead used for switching the scenes mostly. All the other controllers should inject the main controller to gain access to those methods. This way there are no extra injectors.