Added a screen for specifying a server to play on.
The user can go to the screen through the button on the splash screen. There they can input a new server and the client will try to connect to it. Depending on the outcome a feedback message is shown.

The use of MainCtrl.SERVER is now deprecated, since the ServerSelectorCtrl now takes care of what server is used by the client and should therefore be used from now on.