Changed The question retrieval mechanic
we used to make  a request in the game class. Since we want to generalize this class for singleplayer and multiplayer I changed it up a bit so that  we make the request before creating the questions in game class.  Also changed the List activitylist into Arraylists. the question class in commons/game/ is deleted along with it's test. Replaced with Question in models